fixed aes program
