+++ src/stored_procedure/mod.rs
@@ -1,2 +1,22 @@
 pub mod keys;
 pub mod record;
+
+
+/* For testing purposes only
+fn read_nonempty_string_from_user_default(prompt: &str, default: &str) -> String {
+    let mut input = String::from(default);
+    loop {
+        print!("{} [{}]: ", prompt, default);
+        io::stdout().flush().unwrap();
+        io::stdin().read_line(&mut input).unwrap();
+        input = input.trim().to_string();
+        if !input.is_empty() {
+            return input;
+        }
+        println!(
+            "{}",
+            "You must enter a non-empty value. Please try again.".red()
+        );
+    }
+}
+*/
